[easy] Create a data factories package for ImageStack
This just breaks up factories.py into individual files in preparation for more factories being extricated from test code.

Depends on #1174
