+++ starfish/imagestack/test/factories/__init__.py
@@ -1,0 +1,3 @@
+from .from_codebook import create_imagestack_from_codebook
+from .synthetic_stack import synthetic_stack
+from .with_coords import imagestack_with_coords_factory
